refactor: Extract useIsMounted and useSafeState
useSafeState is useful not to call setState when a component is
unmounted
