+++ src/hooks/useIsMounted.jsx
@@ -1,0 +1,14 @@
+import { useRef, useEffect } from 'react'
+
+const useIsMounted = () => {
+  const mounted = useRef()
+  useEffect(() => {
+    mounted.current = true
+    return () => {
+      mounted.current = false
+    }
+  }, [])
+  return mounted
+}
+
+export default useIsMounted
